Improved Customization Added
created a new demo game to begin working on better customizability of how rom are displayed.  The TradeWar4040 demo game is a framework for me to see how far I can push the customization and to be a future show case of how the engine could be used for more then just IF games.